% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dataProcessBayes.R
\name{calculate_priors}
\alias{calculate_priors}
\title{Calculate priors given some data for the run, feature, and error effect}
\usage{
calculate_priors(data, run_only = FALSE)
}
\arguments{
\item{data}{MSstats formatted dataframe}

\item{run_only}{logical. If TRUE, only calculate priors for the run effect}
}
\value{
list of priors for the run, feature, and error effect
}
\description{
Calculate priors given some data for the run, feature, and error effect
}
